% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{predict.survSL.gam}
\alias{predict.survSL.gam}
\title{Prediction functions for generalized additive Cox regression prediction algorithm}
\usage{
\method{predict}{survSL.gam}(object, newX, new.times, ...)
}
\arguments{
\item{object}{Fitted \code{survSL.gam} object.}

\item{newX}{New covariate data.frame for which to obtain predictions.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{...}{Additional ignored arguments.}
}
\value{
Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.
}
\description{
Obtains predicted survivals from a fitted \code{survSL.gam} object.
}
