% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_functions.R
\name{predict.survSuperLearner}
\alias{predict.survSuperLearner}
\title{Prediction function for survival Super Learner}
\usage{
\method{predict}{survSuperLearner}(
  object,
  newdata,
  new.times,
  X = NULL,
  time = NULL,
  event = NULL,
  onlySL = FALSE,
  threshold = 1e-04,
  ...
)
}
\arguments{
\item{object}{Fitted \code{survSuperLearner} object.}

\item{newdata}{\code{r x p} data.frame of new covariates for which to obtain predicted survivals. Must have same names and structure as original covariates \code{X} used to train the SuperLearner.}

\item{new.times}{\code{q x 1} vector of times for which to obtain predicted survivals.}

\item{X}{Original training covariates, which may be needed for obtaining predictions for some learners. Defaults to \code{NULL}.}

\item{time}{Original training follow-up times, which may be needed for obtaining predictions for some learners. Defaults to \code{NULL}.}

\item{event}{Original training event indicators, which may be needed for obtaining predictions for some learners. Defaults to \code{NULL}.}

\item{onlySL}{Logical indicating whether to only fit the learners with non-zero SuperLearner coefficient. Defaults to \code{FALSE}.}

\item{threshold}{Minimum coefficient weight for a learner to be inclued in the super learner prediction. Helps reduce computation time if there are many learners with small weights. Defaults to \code{10e-4}.}
}
\value{
\code{survSuperLearner} returns a named list with the following elements:
\item{call}{The matched call.}
\item{event.libraryNames, cens.libraryNames}{Parsed learner names.}
\item{event.SL.library, cens.SL.library}{Libraries used for fitting.}
\item{event.SL.predict, cens.SL.predict}{\code{m x k} matrices of SuperLearner predicted survival values. Rows index observations in \code{newX}; columns index times in \code{new.times.}}
\item{event.coef, cens.coef}{Fitted SuperLearner coefficients for the model for the conditional survival functions for the event and censoring times, respectively.}
}
\description{
This function predicts the fitted survival Super Learner on new data.
}
\details{
This is the prediction function for \code{\link{survSuperLearner}}.
}
