% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survlistWrappers}
\alias{survlistWrappers}
\title{List wrappers in survSuperLearner}
\usage{
\code{survlistWrappers()}
}
\arguments{
\item{what}{If \code{"both"}, lists both prediction and screening functions; if \code{"survSL"}, lists only prediction functions; if \code{"survscreen"}, lists only screening functions; otherwise all functions in the package.}
}
\value{
Invisible character vector with the requested functions.
}
\description{
This function lists all prediction algorithms and/or screening algorithms built in to the \code{survSuperLearner} package.
}
