% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survscreen.glmnet}
\alias{survscreen.glmnet}
\title{Wrapper function for glmnet screening algorithm}
\usage{
survscreen.glmnet(
  time,
  event,
  X,
  obsWeights,
  alpha = 1,
  minscreen = 2,
  nfolds = 10,
  nlambda = 100,
  ...
)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{obsWeights}{Observation weights.}

\item{alpha}{Penalty exponent for \code{glmnet}. Defaults to 1 (lasso penalty).}

\item{minscreen}{Minimum number of covariates to return. Defaults to 2.}

\item{nfolds}{Number of folds for cross-validation selection of penalty parameter. Defaults to 10.}

\item{nlambda}{Number of penalty parameters to search over. Defaults to 100.}

\item{...}{Additional ignored arguments.}
}
\value{
Logical vector of the same length as the number of columns of \code{X} indicating which variables were included.
}
\description{
This screening algorithm uses the \code{\link[glmnet]{glmnet}} function from the \code{glmnet} package to select covariates.
}
\details{
The penalty parameter is selected using cross-validation via \code{\link[glmnet]{cv.glmnet}}. If this results in fewer than \code{minscreen} covariates, the penalty is increased to include \code{minscreen} covariates.
}
