% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survscreen.marg}
\alias{survscreen.marg}
\title{Wrapper function for marginal Cox regression screening algorithm}
\usage{
survscreen.marg(time, event, X, obsWeights, minscreen = 2, min.p = 0.1, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{obsWeights}{Observation weights.}

\item{minscreen}{Minimum number of covariates to return. Defaults to 2.}

\item{min.p}{Threshold p-value used to decide if a covariate is included. Defaults to 0.1}

\item{...}{Additional ignored arguments.}

\item{alpha}{Penalty exponent for \code{glmnet}. Defaults to 1 (lasso penalty).}
}
\value{
Logical vector of the same length as the number of columns of \code{X} indicating which variables were included.
}
\description{
This screening algorithm uses marginal \code{\link[survival]{coxph}} regressions to select covariates that have significant marginal relationships with the event.
}
\details{
A univariate Cox regression is run for each covariate; those with p-values less than \code{min.p} are included.
}
