% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.coxph}
\alias{survSL.coxph}
\title{Wrapper function for Cox proportional hazards regression prediction algorithm}
\usage{
survSL.coxph(time, event, X, newX, new.times, obsWeights, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{...}{Additional ignored arguments.}
}
\value{
\item{pred}{Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.}
\item{fit}{One-element list including \code{object}, the fitted \code{\link[survival]{coxph}} object.}
}
\description{
This prediciton algorithm uses the partial maximum likelihood estimator of the coefficients and Breslow estimator of the baseline cumulative hazard in the Cox proportional hazards model using the \code{\link[survival]{coxph}} and \code{\link[survival]{survfit}} functions.
}
