% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.expreg}
\alias{survSL.expreg}
\alias{survSL.weibreg}
\alias{survSL.loglogreg}
\title{Wrapper function for zero-inflated parametric survival regression prediction algorithms}
\usage{
survSL.expreg(time, event, X, newX, new.times, obsWeights, ...)

survSL.weibreg(time, event, X, newX, new.times, obsWeights, id, ...)

survSL.loglogreg(time, event, X, newX, new.times, obsWeights, id, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{...}{Additional ignored arguments.}
}
\value{
\item{pred}{Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.}
\item{fit}{Two-element list including \code{reg.object}, the fitted \code{\link[survival]{survreg}} object, and \code{pos.object}, the fitted \code{\link[stats]{glm}} object for the probability that the event was positive (or 1 if no zeroes were detected).}
}
\description{
These prediciton algorithm use the \code{\link[survival]{survreg}} function from the \code{survival} package to estimate parametric survival regressions. See details for specific parametric models.
}
\details{
Parametric regression models assume a particular parametric form for the distribution of the event given covariates, where the covariates contribute in a linear way to one of the parameters of the distribution. Implemented here are currently exponential (\code{survSL.expreg}), Weibull (\code{survSL.weibreg}), and log-logistic (\code{survSL.loglogreg}) regressions.

Note that survival regressions typically assume that the distribution of the event is continuous and strictly positive. Therefore, they will throw errors if there are large discrete components of the observed distribution of the event. Since some survival outcomes have positive mass at zero, we have amended the standard survival regression to include a component at zero, making this a \emph{zero-inflated} regression model. Specifically, if there are observed event times equal to zero, then a preliminary logistic regression is fit to predict the probability that the event time is exactly zero, and the survival regression is fit to the strictly positive part of the distribution.
}
