% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.gam}
\alias{survSL.gam}
\title{Wrapper function for generalized additive Cox regression prediction algorithm}
\usage{
survSL.gam(time, event, X, newX, new.times, cts.num = 5, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{cts.num}{The lower cutoff of unique values at which a covariate should be treated as continuous. Any covariate with number of unique values strictly larger than \code{cts.num} will be treated as continuous and receive a smooth term in the GAM. If \code{X} contains unordered factors with large numbers of unique values, \code{cts.num} should be set to larger than these numbers of unique values, otherwise an error may be thrown.}

\item{...}{Additional ignored arguments.}
}
\value{
\item{pred}{Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.}
\item{fit}{One-element list including \code{object}, the fitted \code{\link[mgcv]{gam}}. object.}
}
\description{
This prediciton algorithm uses the \code{\link[mgcv]{gam}} function from the \code{mgcv} package to estimate a generalized additive Cox proportional hazards regression model. This model generalizes the usual Cox proportional hazards model to allow for an additive combination of smooth and possibly non-linear functions of the continuous covariates.
}
