% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.km}
\alias{survSL.km}
\title{Wrapper function for Kaplan-Meier prediction algorithm}
\usage{
survSL.km(time, event, X, newX, new.times, obsWeights, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{...}{Additional ignored arguments.}
}
\value{
\item{pred}{Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.}
\item{fit}{One-element list including \code{object}, the fitted \code{\link[survival]{survfit}} object.}
}
\description{
This prediciton algorithm ignores all covariates and simply computes the Kaplan-Meier estimator of the marginal survival function of the event as indicated by the right-censored data \code{time} and \code{event} using the \code{\link[survival]{survfit}} function.
}
