% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.pchreg}
\alias{survSL.pchreg}
\title{Wrapper function for piecewise constant hazard regression}
\usage{
survSL.pchreg(time, event, X, newX, new.times, obsWeights, breaks = 4, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{breaks}{Number or numeric vector of breaks to be passed to \code{\link[pch]{pchreg}}.}

\item{...}{Additional ignored arguments.}
}
\description{
These prediciton algorithm use the \code{\link[pch]{pchreg}} function from the \code{pch} package to estimate piecewise constant hazard regressions.
}
