% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_wrappers.R
\name{survSL.template}
\alias{survSL.template}
\title{Wrapper functions for prediction algorithms in survSuperLearner}
\usage{
survSL.template(time, event, X, newX, new.times, obsWeights, id, ...)
}
\arguments{
\item{time}{Observed follow-up time; i.e. minimum of the event and censoring times.}

\item{event}{Observed event indicator; i.e, whether the follow-up time corresponds to an event or censoring.}

\item{X}{Training covariate data.frame.}

\item{newX}{Test covariate data.frame to use for prediction. Should have the same variable names and structure as \code{X}.}

\item{new.times}{Times at which to obtain to obtain the predicted survivals.}

\item{obsWeights}{Observation weights.}

\item{id}{Observation clusters.}

\item{...}{Additional ignored arguments.}
}
\value{
\item{pred}{Matrix of predictions, with the same number of rows as \code{newX} and number of columns equal to the length of \code{new.times}. Rows index new observations, and columns index new times at which the survival was computed.}
\item{fit}{Elements of the model fit. Typically includes \code{fit$object}, the fitted object, but may include additional elements obtained in the fitting process that would be needed for prediction.}
}
\description{
This is a template function for a \code{survSuperLearner} prediction algorithm. You can use this to write your own prediction algorithms.
}
