% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_functions.R
\name{survSuperLearner.control}
\alias{survSuperLearner.control}
\title{Control parameters for the survival Super Learner}
\usage{
survSuperLearner.control(
  initWeightAlg = "survSL.rfsrc",
  initWeight = "censoring",
  max.SL.iter = 20,
  event.t.grid,
  cens.t.grid,
  saveFitLibrary = TRUE
)
}
\arguments{
\item{initWeightAlg}{Algorithm to use for the first iteration of the iterative SuperLearner algorithm. Defaults to \code{survSL.rfsrc}}

\item{initWeight}{Whether to start the iterative SuperLearner by fitting censoring weights (\code{"censoring"}, default) or by fitting event weights (\code{"event"}).}

\item{max.SL.iter}{Maximum iterations of the iterative SuperLearner algorithm. Defaults to 20.}

\item{event.t.grid}{Grid of times to use to approximate the integral in the risk function for the conditional survival function of the event. Defaults to 250 points equally spaced between 0 and the last uncensored follow-up time.}

\item{cens.t.grid}{Grid of times to use to approximate the integral in the risk function for the conditional censoring survival function. Defaults to 250 points equally spaced between 0 and the last censored follow-up time, minus a small constant in order to approximate left-continuous survivals.}

\item{saveFitLibrary}{Logical indicating whether to save the fit library on the full data. Defaults to \code{TRUE}. If \code{FALSE}, cannot obtain predicted values on new data later.}
}
\value{
Returns a named list with control parameters.
}
\description{
This function initiates control parameters for the \code{\link{survSuperLearner}} function.
}
