% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL_functions.R
\name{survSuperLearner.CV.control}
\alias{survSuperLearner.CV.control}
\title{Control parameters for the cross validation steps in survival Super Learner}
\usage{
survSuperLearner.CV.control(
  V = 10L,
  stratifyCV = TRUE,
  shuffle = TRUE,
  validRows = NULL
)
}
\arguments{
\item{V}{Number of cross-validation folds. Defaults to 10.}

\item{stratifyCV}{Logical indicating whether to balance number of observed events across folds. Defaults to \code{TRUE}.}

\item{shuffle}{Logical indicating whether to shuffle the indices, or to simply assign sequentially. Defaults to \code{TRUE}. Should almost always be set to \code{TRUE} unless it is explicitly desired to assign sequentially.}

\item{validRows}{Optional custom list of indices for validation folds.}
}
\value{
Returns a list of length \code{V} with validation indices for each of the folds.
}
\description{
This function initiates control parameters for the cross-validation in \code{\link{survSuperLearner}} function.
}
